Move setup spec.hooks to up layer make it callable for all shims
