+++ crates/wasmtime/src/instance.rs
@@ -163,6 +163,10 @@ impl Instance for Wasi {
 
         debug!("preparing module");
         let spec = load_spec(self.bundle.clone())?;
+
+        debug!("call prehook before the start");
+        oci::setup_prestart_hooks(&spec.hooks())?;
+
         let m = prepare_module(engine.clone(), &spec, stdin, stdout, stderr)
             .map_err(|e| Error::Others(format!("error setting up module: {}", e)))?;
